Move DeclEnum types into their own module
This is in preparation of the Session.nodes property which will merge
information from associated SessionNode objects as well as the Node
objects on the other end of the relationship: Left in the old place,
this would cause a circular import.
