+++ duffy/database/model/__init__.py
@@ -1,9 +1,6 @@
-from .node import VirtualNodeFlavour  # noqa: F401
 from .node import (  # noqa: F401
     Chassis,
     Node,
-    NodeState,
-    NodeType,
     OpenNebulaNode,
     PhysicalNode,
     SeaMicroNode,
